requirements.txt
absl-py==1.4.0
altair==5.0.1
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
decorator==5.1.1
flatbuffers==23.5.26
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keras==2.13.1
lazy_loader==0.3
libclang==16.0.6
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sklearn==0.0.post7
smmap==5.0.0
streamlit==1.24.1
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.18
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
Werkzeug==2.3.6
wrapt==1.15.0
xgboost==1.7.6
zipp==3.16.2
